Use a whitelist to block access from outside
We are now exposing the services using cloudflare tunnels, so we only
allow access from local addresses.

Ideally, we should close ports 80 and 443 entirely, but I have not
managed do that yet, because k3s has added iptables rules for those
ports that are prioritized over the ones created by ufw (and I don't
know enough of iptables to fix that now).